requirements.txt
# Core dependencies for semantic spreadsheet search engine
pandas==2.1.4
openpyxl==3.1.2
numpy==1.24.3

# NLP and semantic processing
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2

# Google AI/Gemini integration
google-generativeai==0.3.2

# Web interface (optional)
flask==3.0.0
flask-cors==4.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2

# Development and testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0

